requirements.txt
boto3==1.34.44
python-dotenv==1.0.0
Pillow==10.2.0